requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-filter==22.1
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.6
